ipaffs_api/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
annotated-types==0.7.0
anthropic==0.49.0
anyio==4.9.0
async-timeout==4.0.3
attrs==25.3.0
banks==2.0.0
blinker==1.9.0
cachelib==0.13.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
dataclasses-json==0.6.7
Deprecated==1.2.18
dirtyjson==1.0.8
distro==1.9.0
et_xmlfile==2.0.0
exceptiongroup==1.2.2
filetype==1.2.0
Flask==3.1.0
flask-cors==6.0.1
Flask-Session==0.8.0
frozenlist==1.5.0
fsspec==2025.3.0
greenlet==3.1.1
griffe==1.6.3
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.9.0
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
langchain==0.3.21
langchain-anthropic==0.3.10
langchain-core==0.3.58
langchain-mcp-adapters==0.1.7
langchain-openai==0.3.16
langchain-text-splitters==0.3.7
langgraph==0.3.20
langgraph-checkpoint==2.0.23
langgraph-prebuilt==0.1.4
langgraph-sdk==0.1.58
langsmith==0.3.18
llama-cloud==0.1.16
llama-cloud-services==0.6.8
llama-index-core==0.12.26
MarkupSafe==3.0.2
marshmallow==3.26.1
mcp==1.10.1
msgspec==0.19.0
multidict==6.2.0
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
numpy==2.2.4
openai==1.77.0
openpyxl==3.1.5
orjson==3.10.15
ormsgpack==1.9.0
packaging==24.2
pandas==2.2.3
pillow==11.1.0
propcache==0.3.1
psycopg2-binary==2.9.10
pydantic==2.10.6
pydantic-settings==2.10.1
pydantic_core==2.27.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2025.1
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rpds-py==0.26.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.39
sse-starlette==2.4.1
starlette==0.47.1
tenacity==9.0.0
tiktoken==0.9.0
tqdm==4.67.1
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
uvicorn==0.35.0
Werkzeug==3.1.3
wrapt==1.17.2
xxhash==3.5.0
yarl==1.18.3
zstandard==0.23.0
